Move agujeronegro utility
