+++ react/codicon/src/pages/blackhole.jsx
@@ -1,5 +1,5 @@
 import "../css/blackhole.css"
-import "../css/agujeronegro.js"
+import "../utils/agujeronegro.js"
 import { useEffect } from "react";
 import cerrada from "../img/cerrada.png";
 import Stars from "./stars";
